chore: split up main.js code into methods & helpers
