+++ elements/tour/src/methods/tour/index.js
@@ -1,0 +1,3 @@
+export { handleMessage } from "./handle-message.js";
+export { initializeDriver } from "./initialize-driver.js";
+export { showTourOrNot } from "./show-tour.js";
